backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic
python-dotenv

# Firebase & Auth
firebase-admin

# LangChain & AI
langchain
langchain-google-genai
langchain-community
google-genai

# Vector Store & Embeddings
faiss-cpu
sentence-transformers

# Data Processing
pandas
numpy

# Web Scraping & Search
duckduckgo-search
beautifulsoup4
requests

# Document Loaders
pypdf
python-docx

# Utilities
python-multipart
aiofiles
